Changes to Parser and new modules Frame and Core
Frame module implements a data structure for representing graphs.

This will be used for frame and model implementations. Some basic
graph property checking functions have been implemented as they will
be the core part for defining logics in the program.

Models will have the same semantics (worlds, transitions and tags), but in order

Frame module implements a data structure for representing graphs.
This will be used for frame and model implementations. Some basic
graph property checking functions have been implemented as they will
be the core part for defining logics in the program.

Models will have the same semantics (worlds, transitions and tags), but in order
to use frames, the latter have been abstracted into their own data type and so
a model is composed now of a frame and the tagging map.

examples/base.mll contains a general idea of what I expect to do.
Probably will add more things, but the syntax is pretty much a closed

Updated parser to parse set statements, used to modify the model
that the user works with. There are 3 kinds of set stmts. which are
set worlds, set transition and set tag.

The parser also parses now many representations for some
operators, like "and" and "&&" for the conjunction. Substitution
of formulas has been also implemented with help of the sub function.
The latter uses the monad instance for formulas.

Core.hs contains definitions used by Parser.hs and Modal.hs