Many monads involved in evaluation.
State monads for statements execution, Except monads
for error handling and Readers for expression evaluation.

Some changes done to syntax. `set worlds` is no longer used,
instead frames are built explicitly from the edges specification,
which is now written as `set frame` instead of `set transitions`

Also I have no idea on how to implement formula substitutions
properly. Implementing it inside the parser is probably the best,
but it requires the parser to keep a state on the actual definitions.
On the other hand, doing it in the runtime means that definitions
are not constant, but rather they feed on the state at the time
the expression is being evaluated, most likely not what the user
would expect. This is however the most easy thing to do (I think).