Implemented global identifiers and substitutions.
Global identifiers lookup is implemented along the parser. The parser
actually parses a data type Scheme, which then given a lookup table
resolves global definitions.

Two data types were needed for substitution implementation. The first
one is LitFormula, thought as a literal and more concrete translation
of the abstract syntax of the language. Formula is the abstract
representation intended for evaluation. toFormula function solves
substitutions and translates the data types as these are equivalent
when subs are performed.

I could not figure out how to implement the idea of performing
substitutions inside the parser. However this way the parser is simpler.

Common.hs contains mostly data types used by all modules.