Pretty printing implemented.
Of course it's not complete, but it already shows formulas
with unicode characters and colors. Top and Bottom are displayed
with green and red respectively, atoms are assigned colors from
the remaining ones.

Next step is implementing logic choice and axiom verification.
Also, a verbose mode where the program displays its evaluation
trace could be really nice human interpretation.