Trace implementation.
Evaluator now returns a trace of evaluation. This way, on top of
evaluating a formula, the evaluator actually contains the evaluation
of every subformula, and it can be used for an enchanced output mode.

A basic pretty printing has been implemented only for single traces
and not yet for model traces.