Modules reordered, main partially done.
Fixed some issues with cyclic imports in modules. Some definitions
where moved to avoid this and improve abstraction.

main added, although its not finished so its not compiling in this
version.

eval function and traces where modified a little bit. The idea
was to abstract the three types of evaluation traces in a Trace class,
but I havent found the right way to do it so far.